display turnament tables
