+++ src/Volley/FaceBundle/Resources/views/Stat/index.html.twig
@@ -1,0 +1,1 @@
+Hello {{ name }}!
