+++ src/Volley/FaceBundle/Resources/views/Stat/tournament.html.twig
@@ -1,0 +1,94 @@
+{% extends '::layout.html.twig' %}
+
+{% block content %}
+
+    <div class="section">
+        <div class="container">
+            <div class="row">
+                <h1>{{ tournament.name }} - {{ season.name }}</h1>
+            </div>
+            <div class="row">
+                <table class="table table-striped table-hover table-condensed">
+                    <thead>
+                    <th>Місце</th>
+                    <th>Команда</th>
+                    <th>Матчі</th>
+                    <th>Перемоги</th>
+                    <th>Поразки</th>
+                    <th>3-0</th>
+                    <th>3-1</th>
+                    <th>3-2</th>
+                    <th>2-3</th>
+                    <th>1-3</th>
+                    <th>0-3</th>
+                    <th>Партії</th>
+                    <th>М'ячі</th>
+                    <th>Очки</th>
+                    </thead>
+                    {% for row in table %}
+                        <tr>
+                            <td>{{ loop.index }}</td>
+                            {#<td><a href="{{ path('volley_face_stat_team',{'team_id':row['team'].id}) }}">{{ row['team'].name }}</a></td>#}
+                            <td>{{ row['team'].name }}</td>
+                            <td>{{ row.games }}</td>
+                            <td>{{ row.win }}</td>
+                            <td>{{ row.loss }}</td>
+                            <td>{{ row.score30 }}</td>
+                            <td>{{ row.score31 }}</td>
+                            <td>{{ row.score32 }}</td>
+                            <td>{{ row.score23 }}</td>
+                            <td>{{ row.score13 }}</td>
+                            <td>{{ row.score03 }}</td>
+                            <td>{{ row.win_sets }}/{{ row.loss_sets }}</td>
+                            <td>{{ row.win_points }}/{{ row.loss_points }}</td>
+                            <td>{{ row.points }}</td>
+                        </tr>
+                    {% endfor %}
+                </table>
+            </div>
+            <div class="row">
+                <table class="table table-striped table-hover table-condensed">
+                    <thead>
+                    <tr>
+                        <td>#</td>
+                        <td>Дата</td>
+                        <td>Час</td>
+                        <td>Команда А</td>
+                        <td>Команда Б</td>
+                        <td>Результат</td>
+                    </tr>
+                    </thead>
+                    <tfoot></tfoot>
+                    <tbody>
+                    {% for round in season.rounds %}
+                        {% for tour in round.tours %}
+                            <tr>
+                                <td colspan="100" class="text-center"><strong>{{ tour.name }}</strong></td>
+                            </tr>
+                            {% for game in tour.games %}
+                                {% set winnerTeamA = (game.scoreSetHome > game.scoreSetAway) %}
+                                {% set winnerTeamB = (game.scoreSetHome < game.scoreSetAway) %}
+                                <tr>
+                                    <td>{{ game.number }}</td>
+                                    <td>{% if game.date is not null %} {{ game.date|date("d/m") }} {% endif %}</td>
+                                    <td>{% if game.date is not null %} {{ game.date|date("H:i") }} {% endif %}</td>
+                                    <td>{{ winnerTeamA ? "<strong>" : "" }}
+                                        {#<a href="{{ path('volley_face_stat_team',{'team_id':game.homeTeam.id}) }}"> {{ game.homeTeam.name }} </a>{{ winnerTeamA ? "</strong>" : "" }}#}
+                                        {{ game.homeTeam.name }} {{ winnerTeamA ? "</strong>" : "" }}
+                                    </td>
+                                    <td>{{ winnerTeamB ? "<strong>" : "" }}
+                                        {#<a href="{{ path('volley_face_stat_team',{'team_id':game.awayTeam.id}) }}"> {{ game.awayTeam.name }} </a>{{ winnerTeamB ? "</strong>" : "" }}#}
+                                        {{ game.awayTeam.name }} {{ winnerTeamB ? "</strong>" : "" }}
+                                    </td>
+                                    <td>{{ game.score }}</td>
+                                </tr>
+                            {% endfor %}
+                        {% endfor %}
+                    {% endfor %}
+                    </tbody>
+                </table>
+            </div>
+        </div>
+    </div>
+
+{% endblock %}
