+++ src/Volley/StatBundle/Controller/TournamentController.php
@@ -249,6 +249,8 @@ class TournamentController extends Controller
         ;
     }
 
+
+
     /**
      * * Deletes a Tournament entity.
      *
@@ -260,105 +262,7 @@ class TournamentController extends Controller
      */
     public function tableAction($seasonId = 1, $tournamentId = 1)
     {
-        $em = $this->getDoctrine()->getManager();
-
-        /** @var Season $season */
-        $season = $em->getRepository('VolleyStatBundle:Season')->find($seasonId);
-
-        /** @var Tournament  $tournament */
-        $tournament = $em->getRepository('VolleyStatBundle:Tournament')->find($tournamentId);
-
-        $teams = $season->getTeams();
-
-        $games = $season->getGames();
-
-        $table = [];
-        foreach ($teams as $team) {
-            $table[$team->getID()] = ['team' => $team, 'points' => 0, 'games' => 0, 'win' => 0, 'loss' => 0, 'win_sets' => 0, 'loss_sets' => 0, 'win_points' => 0, 'loss_points' => 0];
-        }
-
-        /**
-         * @var Game $game
-         */
-        foreach ($games as $game) {
-
-            if (!$game->getPlayed())
-                continue;
-
-            if ($game->getHomeTeam()) {
-                $homeTeamId = $game->getHomeTeam()->getId();
-                $table[$homeTeamId]['games'] += 1;
-            }
-            else
-                $homeTeamId = 0;
-            if ($game->getAwayTeam()) {
-                $awayTeamId = $game->getAwayTeam()->getId();
-                $table[$awayTeamId]['games'] += 1;
-            }
-            else
-                $awayTeamId = 0;
-
-            $homeTeamSets = $game->getScoreSetHome();
-            $awayTeamSets = $game->getScoreSetAway();
-
-            if ($homeTeamSets > $awayTeamSets) {
-                if ($homeTeamId) $table[$homeTeamId]['win'] += 1;
-                if ($homeTeamId) $table[$homeTeamId]['win_sets'] += $homeTeamSets;
-                if ($awayTeamId) $table[$awayTeamId]['loss'] += 1;
-                if ($awayTeamId) $table[$awayTeamId]['loss_sets'] += $awayTeamSets;
-                if ($homeTeamSets - $awayTeamSets >= 2) {
-                    if ($homeTeamId) $table[$homeTeamId]['points'] += 3;
-                    if ($awayTeamId) $table[$awayTeamId]['points'] += 0;
-                } else {
-                    if ($homeTeamId) $table[$homeTeamId]['points'] += 2;
-                    if ($awayTeamId) $table[$awayTeamId]['points'] += 1;
-                }
-            } else {
-                if ($homeTeamId) $table[$homeTeamId]['loss'] += 1;
-                if ($homeTeamId) $table[$homeTeamId]['loss_sets'] += $homeTeamSets;
-                if ($awayTeamId) $table[$awayTeamId]['win'] += 1;
-                if ($awayTeamId) $table[$awayTeamId]['win_sets'] += $awayTeamSets;
-                if ($awayTeamSets - $homeTeamSets >= 2) {
-                    if ($homeTeamId) $table[$homeTeamId]['points'] += 0;
-                    if ($awayTeamId) $table[$awayTeamId]['points'] += 3;
-                } else {
-                    if ($homeTeamId) $table[$homeTeamId]['points'] += 1;
-                    if ($awayTeamId) $table[$awayTeamId]['points'] += 2;
-                }
-            }
-
-            /** @var GameSet $set */
-            foreach ($game->getSets() as $set) {
-                if ($homeTeamId) $table[$homeTeamId]['win_points'] += $set->getScoreSetHome();
-                if ($homeTeamId) $table[$homeTeamId]['loss_points'] += $set->getScoreSetAway();
-                if ($awayTeamId) $table[$awayTeamId]['win_points'] += $set->getScoreSetAway();
-                if ($awayTeamId) $table[$awayTeamId]['loss_points'] += $set->getScoreSetHome();
-            }
-        }
-
-        foreach ($table as &$row) {
-            $row['k'] = $row['points']*1000000 + ($row['loss_sets']?$row['win_sets']/$row['loss_sets']:$row['win_sets'])*1000 + ($row['loss_points']?$row['win_points']/$row['loss_points']:$row['win_points']);
-        }
-
-//        usort($table, function ($a, $b) {
-//            return strcmp($b['points'] + ($b['loss_sets']?$b['win_sets']/$b['loss_sets']:$b['win_sets'])/1000 + ($b['loss_points']?$b['win_points']/$b['loss_points']:$b['win_points'])/1000000,
-//                $a['points'] + ($a['loss_sets']?$a['win_sets']/$a['loss_sets']:$a['win_sets'])/1000 + ($a['loss_points']?$a['win_points']/$a['loss_points']:$a['win_points'])/1000000);
-//        });
-        usort($table, function ($a, $b) {
-            return $this->cmp($b['k'], $a['k']);
-        });
-
-        return $this->render('VolleyStatBundle:Tournament:table.html.twig', array(
-            'season' => $season,
-            'tournament' => $tournament,
-            'table' => $table
-        ));
+        return $this->render('VolleyStatBundle:Tournament:table.html.twig', $this->get('volley_stat.tournament.manager')->getTournamentData($seasonId, $tournamentId));
     }
 
-    private function cmp($a, $b) {
-        if ($a == $b) {
-            return 0;
-        }
-        return ($a < $b) ? -1 : 1;
-    }
 }
