+++ src/Volley/StatBundle/Resources/views/Tournament/table.html.twig
@@ -1,6 +1,6 @@
 <div class="row">
     <div class="col-lg-12">
-        <h2>{{ tournament.name }}</h2>
+        <h2><a href="{{ path('stat_tournament_page', {'season_id': 1, 'tournament_id': 1}) }}">{{ tournament.name }}</a></h2>
         <table class="table table-hover table-condensed">
             <thead class="text-center">
             <th title="Місце">#</th>
